Add new options to NnfStorageProfile
CapacityScalingFactor: An option to increase the capacity in the directive breakdown
from what the user requested on the #dw line. This can be used to make the usable
file system space more closely match the user's request. An admin must set this to a
value that makes sense with the block device and file system options used. This is
separately configurable for each file system type.

StorageLabels: A set of labels to add to the labels constraint in the directive breakdown.
This is separately configurable for each file system type and each Lustre target type.

I also changed the #dw create_persistent directive to no longer require the "capacity"
argument when it's being used to create standalone MGTs.

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>