Check NnfNodeBlockStorage is ready before reading device list
We've had situations where the NnfNodeStorage reads the NnfNodeBlockStorage device list
and gets an empty list of devices. This could be caused by a stale NnfNodeBlockStorage
resource in the cache. The status.ready flag in the NnfNodeBlockStorage won't be set
until everything is filled in. Check that flag before reading any values.

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>