Add env var for optional pause during DataIn
Data Movement nightly tests routinely hit this issue:
https://github.com/NearNodeFlash/NearNodeFlash.github.io/issues/161.

This appears to only affect a specific version of lustre/toss on our system until
we upgrade.

This change adds an optional workaround for to add a pause to let lustre
settle after mounting. 1s appears to be all that is needed.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>