Add PostMount/PreUnmount Support to lustre; Rename PostActivate/PreDeactivate for non-lustre
We currently do not have a way to perform actions (e.g. `lfs setstripe`)
on lustre filesystems from the client side. For situations like
`DataIn`, there is no way to prepare the lustre filesystem prior to data
movement.

This change adds PostMount and PreUnmount command lines to the
NnfStorageProfile to allow commands to be run in those situations.

- For lustre filesystems, PostMount and PreUnmount have been added in
  addition to the existing PostActivate/PreDeactivate commands
- PostActivate/PreDeactivate are performed server-side
- PostMount/PreUnmount are performed client-side
- For XFS/GFS2 filesystems, PostActivate/PreDeactivate have been
  renamed to PostMount/PreUnmount

For lustre, multiple NnfNodeStorages are created for each OST, MDT, and
MGT. PostMount should only happen once, so OST0 is what is used. Before
the filesystem can be mounted to run the commands, we need ensure that
all other NnfNodeStorages are ready. Once that happens, OST0 can then be
created and then the NnfNodeStorage controller can run the PostMount
commands. The opposite logic applies in the PreUnmount case where OST0
is now deleted first and performs the PreUnmount commands.

For XFS/GFS2, there is no issue of ordering.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>