Enable mock filesystems in KIND
On the docker host, the /tmp/nnf dir is expected to be mounted into each
docker container as /mnt/nnf. This should be specified as an extraMounts in
the KIND config file.

The k8s pods will add volume mounts for /mnt/nnf.

When using KIND, new mock devices will be represented as directories in
/mnt/nnf. New filesystems will be represented as directories in /mnt/nnf,
containing per-rabbit symlinks back to the mock "device" directory.

For a user container to use these mock filesystems, their pod must have a
volume mount for the mock filesystem and for the mock device directory, where
the symlink is pointing.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>