PostMount Env Variables (e.g. NUM_OSTS)
There is no way to get the number of OSTs, etc when using `PostMount`
commands. For instance, when setting the striping.

This change adds the following environment variables for use in the
`NnfStorageProfiles` when using `*CmdLines`:

- NUM_MDTS
- NUM_MGTS
- NUM_MGTMDTS
- NUM_OSTS
- NUM_NNFNODES

To support this, a list of nodes for each component type has been added
to the status of `NnfStorage`, which in turn is copied to the
`NnfNodeStorage` resource's spec. This info can then be turned into the
environment variables for use when running the commands.

The `.nnf-servers.json` file created to store this information on the
compute node has been updated to this new structure.

An example with 8 OSTs and 2 MGTMDTs on 1 rabbit:

```json
{
  "mdt": [],
  "mgt": [],
  "mgtmdt": [
    "rabbit-node-1",
    "rabbit-node-1"
  ],
  "nnfNode": [
    "rabbit-node-1"
  ],
  "ost": [
    "rabbit-node-1",
    "rabbit-node-1",
    "rabbit-node-1",
    "rabbit-node-1",
    "rabbit-node-1",
    "rabbit-node-1",
    "rabbit-node-1",
    "rabbit-node-1"
  ]
}
```

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>