Fix NnfAccess ClientMount count
This was a workaround to account for creating multiple OSTs in an
unorthodox manner in the Servers resource. Flux won't create multiple
allocation sets on a single rabbit, but rather use the count when there
are multiple allocations on a single rabbit.

This workaround causes a bug and is not needed.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>