Add LV tags for compute node hostname
This commit adds a method for providing variable substitution to commands run on the Rabbits
from the upper layers. This is used by the NnfSystemStorage controller to add a variable that
expands to the host name of one compute node per Rabbit allocation. This allows the NnfStorageProfile
used by the NnfSystemStorage to use the $COMPUTE_HOSTNAME variable in the lvcreate command to
add a tag to each LV corresponding to a compute node hostname.

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>