Provide a per-Workflow token (#453)
A per-Workflow signing key and token will be created for user containers when
the user requests it on their "#DW" directive. They request this by adding a
"requires=copy-offload" or "requires=user-container-auth" keyword on the jobdw
or persistentdw directive. The Secrets containing the token and its key will be
deleted when the Workflow is deleted.

User container Deployment resources created by nnf-sos will mount Secrets
containing TLS cert&key and token key, if requested by the user on their
"#DW" directive. These allow the user container to contain a server that is
fully locked down, accepting commands from only those computes involved with
the Workflow. The copy-offload server provided by the nnf-dm repo utilizes
these pieces, and it is left to individual users to write their own servers to
utilize these pieces.

The user-container TLS cert&key are created separately. Currently this is
a tool in the nnf-dm repo.

Define "user-container-auth" as valid in the "requires" keyword in
config/dws/nnf-ruleset.yaml for any jobdw or persistentdw.

Update the WLM Role to allow it to read Secrets. The WLM will use this to
get the per-Workflow token from the Workflow, specified in
.status.workflowToken, so it can make the token available to the compute
application.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>