Run commands on MGT after Lustre targets are activated (#458)
* Run commands on MGT after Lustre targets are activated

This commit adds the ability to run commands on an MGT after all the Lustre targets
have been activated. This works for MGTs created with a workflow and MGTs from a
pool.

The commands are specified in the storageprofile.data.lustreStorage.preMountMGTCommands field.

This commit also coordinates the creation of MDTs and OSTs to ensure that all MDTs are created
first.

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>

* review comments

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>

* fix MGT command function comments

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>

* Add log message for external mgt

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>

---------

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>