Containers: Open numPorts host ports for each container
The current implementation did not account for multiple containers in a
PodSpec. This lead to an issue with mapping the same port across multiple
containers, which will not work.

This change counts the number of containers in the spec and opens
numPorts ports for each of those containers. For non-mpi containers,
this would happen for each container. For mpi containers, this only
happens on the Launcher containers - the workers will only be accessible
via mpirun through the launcher.

Like before, the port numbers are provided via the NNF_CONTAINER_PORTS
environment variable.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>