DM Profile: Add $VAR checks for commands in webhook
Make sure that the data movement commands contain the required $VAR in
the command string.

Only do this when the expected command is found (e.g. dcp, mkdir) first.
This allows for open configuration of the commands for testing/debugging
purposes. For example, using `sleep 5` instead of an actual `dcp`
command. Otherwise the validation would fail.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>