Reduce CRD Size by Replacing core.PodSpecs with Slimmed Nnf* Versions
The CRD size has grown too large due to the use of core.PodSpecs and
mpi-operator's MPIJobSpec, making it unwieldy and difficult to manage.

To address this, slimmed-down versions of PodSpec and Container have
been introduced, which are now used for the container profile's Spec and
MPISpec fields.

Additionally:
- A `copyOffload` flag has been added to the MPISpec to simplify
  defining the Copy Offload container profile by automating the correct
  service account usage.
- The `slotsPerWorker` field has been added to the NnfMPISpec to allow
  adjustment of slots in user container jobs. Copy Offload defers to the
  Data Movement profile's slot configuration.

Changes to the NnfDataMovementManager will follow in a subsequent PR.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>