Fix StorageGroup deletion for Kind environment
The StorageGroups for compute nodes in the Kind environment is faked out. In the NnfNodeBlockStorage
Status section we used a fake StorageGroup name as a reminder that this is the case. With the new
code that tracks when StorageGroups are fully created and fully deleted more carefully, this was
causing problems. Use the real StorageGroup name so that the Kind environment is consistent with
real systems.

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>