User Containers: Send `/shutdown` POST to long-running containers (#499)
Any long-running user container (e.g., copy offload) that does not
shut down on its own will get stuck in PostRun until the
`postRunTimeoutSeconds` timeout is hit. This results in an error
since the containers did not finish in the PostRun state before
the timeout.

This change adds a shutdown request to the PostRun logic for user containers.
The long-running container **must** implement an HTTPS server that
accepts a `POST` to the `/shutdown` endpoint. This allows the user
container to shut down gracefully and return success, allowing PostRun
to proceed.

If the shutdown request fails, the workflow continues. In fact, we expect
it to eventually fail since the user container will shut down and no longer
respond. Any errors are logged and do not affect the workflow outcome.
If there are issues sending the shutdown request, they are logged and the
`postRunTimeoutSeconds` timeout will eventually be hit.

If `postRunTimeoutSeconds` is set to 0, PostRun does not wait for a
result; it simply ignores the container result and transitions to
Teardown, where the containers are cleaned up. This can be used for
long-running containers that do not implement a shutdown, allowing
shutdown to be forced by transitioning to Teardown.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>