containers: fix homedir permissions
We dynamically change the UID of `mpiuser` inside of a container.
However, the homedir still has the original GID permissions that were
created when the image is built. This leads to permissions issues
when the container tries to do something in /home/mpiuser (e.g. mkdir).

This adds a mount for /home/$USER. In the passwd InitContainer, the
contents of the original homedir are copied into the mount and then
chown is used to set permissions. This new homedir is then mounted
into the main container.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>