Automate RAID rebuild for LVM and Zpool (#520)
* Automate RAID rebuild for LVM and Zpool

Check the health of the RAID device after it's been created to see if it needs to be repaired.
If the RAID device is unhealthy, check whether there are any new NVMe namespaces to replace the
failed device. This process differs depending on the RAID device type:

zpool:
 Use the "zpool replace" command

lvm:
 Add a PV for the new NVMe namespace
 Extend the existing VG
 Repair the LV
 Reduce the exisitng VG

Currently, the lvm code works, but the setup time for an LVM device with RAID enabled can take
a long time for large allocations due to the need for an initial scrub.

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>

* Use zpool replace command from storage profile

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>

* Address comments

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>

* review comments

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>

---------

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>