Move NnfStorageProfile UserCommand fields
This commit makes changes to the NnfStorageProfile to group things in an easier
to understand way. It also creates a shared struct for XFS, GFS2, and raw parameters
in the NnfStorageProfile so those three allocation types don't have to be maintained
separately.

This commit also logs the output of any user commands specified in the NnfStorageProfile.

I also removed the shared volume group from XFS and Raw allocations since lvmlockd isn't
enabled by default anymore.

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>